refactor(task_executor): extract gates + non-impl path; bring mod.rs under 800-line cap
Closes #820.

Changes:
- New `task_executor/gates.rs`: defines a shared `MANUAL_RESOLUTION_REQUIRED`
  constant and moves the resumed-PR conflict gate, rebase-conflict failure
  helper, and review-entry decision logic out of `mod.rs`. Both the gate
  emission sites and the `intake/github_issues.rs` poller now reference the
  same constant, so a wording change cannot silently break the dispatch
  contract.
- New `task_executor/non_implementation.rs`: hosts `run_non_implementation_task`
  (used by Review/Planner kinds), previously inlined in `mod.rs`.
- New `task_executor/gates_tests.rs` and `task_executor/mod_tests.rs`: move
  the existing gate and `mod.rs` tests behind `#[path]` attributes, matching
  the pattern used by `pr_detection`, `triage_pipeline`, and `helpers`.
- `mod.rs` shrinks from 1752 to 725 lines.

Behavior is unchanged: error wording, persisted round results, event log
shapes, and the Gate A/B/C control flow are preserved verbatim.
