+++ crates/harness-server/src/task_executor/mod.rs
@@ -1,23 +1,23 @@
 pub(crate) mod agent_review;
 pub(crate) mod conflict_resolver;
+pub(crate) mod gates;
 pub(crate) mod helpers;
 pub(crate) mod implement_pipeline;
+pub(crate) mod non_implementation;
 pub(crate) mod pr_detection;
 pub(crate) mod review_loop;
 pub(crate) mod triage_pipeline;
 pub(crate) mod turn_lifecycle;
 
 use crate::task_runner::{
-    mutate_and_persist, CreateTaskRequest, RoundResult, TaskId, TaskKind, TaskStatus, TaskStore,
+    mutate_and_persist, CreateTaskRequest, TaskId, TaskKind, TaskStatus, TaskStore,
 };
 use anyhow::Context;
-use harness_core::agent::{AgentRequest, CodeAgent};
+use harness_core::agent::CodeAgent;
 use harness_core::config::agents::CapabilityProfile;
-use harness_core::tool_isolation::validate_tool_usage;
 use harness_core::{config::project::load_project_config, lang_detect, prompts};
 use std::collections::HashMap;
 
-use chrono::Utc;
 use helpers::update_status;
 
 /// Extract tool list from a capability profile, returning an error if the
@@ -26,7 +26,7 @@ use helpers::update_status;
 /// per U-23 (no silent capability downgrade).
 // Re-export so existing call sites in handlers/ don't need updating.
 pub(crate) use turn_lifecycle::run_turn_lifecycle;
-fn restricted_tools(profile: CapabilityProfile) -> anyhow::Result<Vec<String>> {
+pub(crate) fn restricted_tools(profile: CapabilityProfile) -> anyhow::Result<Vec<String>> {
     profile.tools().ok_or_else(|| {
         anyhow::anyhow!(
             "capability profile {:?} returned None from tools() — misconfiguration",
@@ -77,101 +77,6 @@ pub(crate) struct TaskContext {
 
 fn should_run_issue_triage(skip_triage: bool, has_existing_pr: bool) -> bool {
     !skip_triage && !has_existing_pr
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum ReviewEntryDecision {
-    Proceed { rebase_pushed: bool },
-    FailConflict { error: String, paths_csv: String },
-}
-
-fn review_entry_decision(
-    pr_num: u64,
-    review_prep: Option<&prompts::PrReviewPrepOutcome>,
-) -> ReviewEntryDecision {
-    match review_prep {
-        Some(prompts::PrReviewPrepOutcome::RebasePushed) => ReviewEntryDecision::Proceed {
-            rebase_pushed: true,
-        },
-        Some(prompts::PrReviewPrepOutcome::RebaseSkipped) | None => ReviewEntryDecision::Proceed {
-            rebase_pushed: false,
-        },
-        Some(prompts::PrReviewPrepOutcome::RebaseConflict { paths }) => {
-            let paths_csv = if paths.is_empty() {
-                "unknown paths".to_string()
-            } else {
-                paths.join(", ")
-            };
-            ReviewEntryDecision::FailConflict {
-                error: format!(
-                    "PR #{pr_num} has rebase conflicts: {paths_csv}; manual resolution required"
-                ),
-                paths_csv,
-            }
-        }
-    }
-}
-
-fn review_prep_from_rounds(rounds: &[RoundResult]) -> Option<prompts::PrReviewPrepOutcome> {
-    rounds
-        .iter()
-        .rev()
-        .find(|round| round.action == "implement")
-        .and_then(|round| match round.result.as_str() {
-            "rebase_pushed" => Some(prompts::PrReviewPrepOutcome::RebasePushed),
-            "rebase_skipped" => Some(prompts::PrReviewPrepOutcome::RebaseSkipped),
-            "rebase_conflict" => round
-                .detail
-                .as_deref()
-                .and_then(prompts::parse_pr_review_prep_outcome)
-                .or_else(|| {
-                    Some(prompts::PrReviewPrepOutcome::RebaseConflict { paths: Vec::new() })
-                }),
-            _ => None,
-        })
-}
-
-fn review_prep_from_checkpoint_phase(last_phase: &str) -> Option<prompts::PrReviewPrepOutcome> {
-    match last_phase {
-        "rebase_pushed" => Some(prompts::PrReviewPrepOutcome::RebasePushed),
-        "rebase_skipped" => Some(prompts::PrReviewPrepOutcome::RebaseSkipped),
-        "rebase_conflict" => {
-            Some(prompts::PrReviewPrepOutcome::RebaseConflict { paths: Vec::new() })
-        }
-        _ => None,
-    }
-}
-
-async fn fail_rebase_conflict(
-    store: &TaskStore,
-    task_id: &TaskId,
-    events: &Arc<harness_observe::event_store::EventStore>,
-    pr_num: u64,
-    error: String,
-    paths_csv: String,
-) -> anyhow::Result<()> {
-    mutate_and_persist(store, task_id, |s| {
-        s.status = TaskStatus::Failed;
-        s.error = Some(error.clone());
-    })
-    .await?;
-    store.log_event(crate::event_replay::TaskEvent::Failed {
-        task_id: task_id.0.clone(),
-        ts: crate::event_replay::now_ts(),
-        reason: error,
-    });
-    let mut event = harness_core::types::Event::new(
-        harness_core::types::SessionId::new(),
-        "pr_rebase_conflict",
-        "task_runner",
-        harness_core::types::Decision::Block,
-    );
-    event.reason = Some(paths_csv);
-    event.detail = Some(format!("task_id={} pr={pr_num}", task_id.as_str()));
-    if let Err(err) = events.log(&event).await {
-        tracing::warn!("failed to log pr_rebase_conflict event: {err}");
-    }
-    Ok(())
 }
 
 /// Run the project's test commands as a hard gate before accepting LGTM.
@@ -257,503 +162,6 @@ async fn run_test_gate(
     Ok(())
 }
 
-#[allow(clippy::too_many_arguments)]
-async fn run_non_implementation_task(
-    store: &TaskStore,
-    task_id: &TaskId,
-    task_kind: TaskKind,
-    agent: &dyn CodeAgent,
-    req: &CreateTaskRequest,
-    project: &std::path::Path,
-    server_config: &harness_core::config::HarnessConfig,
-    interceptors: &Arc<Vec<Arc<dyn harness_core::interceptor::TurnInterceptor>>>,
-    events: &Arc<harness_observe::event_store::EventStore>,
-    skills: &Arc<RwLock<harness_skills::store::SkillStore>>,
-    cargo_env: &HashMap<String, String>,
-    turn_timeout: Duration,
-    effective_max_turns: Option<u32>,
-    turns_used: &mut u32,
-    turns_used_acc: &mut u32,
-    task_start: Instant,
-) -> anyhow::Result<()> {
-    let Some(system_input) = req.system_input.as_ref() else {
-        anyhow::bail!(
-            "{} task is missing restart-safe input metadata",
-            task_kind.as_ref()
-        );
-    };
-
-    update_status(store, task_id, task_kind.execution_status(), 1).await?;
-
-    let mut prompt = implement_pipeline::prepend_constitution(
-        system_input.prompt().to_string(),
-        server_config.server.constitution_enabled,
-    );
-    let skill_match_prompt = prompt.clone();
-    let skill_additions = helpers::inject_skills_into_prompt(skills, &skill_match_prompt).await;
-    prompt = helpers::inject_project_context_into_prompt(project, prompt);
-    if !skill_additions.is_empty() {
-        prompt.push_str(&skill_additions);
-    }
-    let context_items = helpers::collect_context_items(skills, project, &skill_match_prompt).await;
-    let allowed_tools = Some(restricted_tools(CapabilityProfile::Standard)?);
-    if let Some(note) = CapabilityProfile::Standard.prompt_note() {
-        prompt = format!("{note}\n\n{prompt}");
-    }
-
-    let prompt_built_at = Utc::now();
-    let initial_req = AgentRequest {
-        prompt,
-        project_root: project.to_path_buf(),
-        context: context_items,
-        max_budget_usd: req.max_budget_usd,
-        execution_phase: Some(harness_core::types::ExecutionPhase::Planning),
-        allowed_tools: allowed_tools.clone(),
-        env_vars: cargo_env.clone(),
-        ..Default::default()
-    };
-    let first_req = helpers::run_pre_execute(interceptors, initial_req).await?;
-    let max_validation_retries: u32 = interceptors
-        .iter()
-        .filter_map(|i| i.max_validation_retries())
-        .max()
-        .unwrap_or(2);
-    let mut validation_attempt = 0u32;
-    let mut turn_req = first_req.clone();
-
-    let resp = loop {
-        if let Some(max) = effective_max_turns {
-            if *turns_used >= max {
-                anyhow::bail!(
-                    "Turn budget exhausted: used {} of {} allowed turns",
-                    turns_used,
-                    max
-                );
-            }
-        }
-        let agent_started_at = Utc::now();
-        let raw = tokio::time::timeout(
-            turn_timeout,
-            helpers::run_agent_streaming(
-                agent,
-                turn_req.clone(),
-                task_id,
-                store,
-                1,
-                prompt_built_at,
-                agent_started_at,
-            ),
-        )
-        .await;
-        *turns_used += 1;
-        *turns_used_acc = *turns_used;
-        match raw {
-            Ok(Ok(success)) => {
-                let response = &success.response;
-                let turn_tools = turn_req.allowed_tools.as_deref().unwrap_or(&[]);
-                let tool_violations = validate_tool_usage(&response.output, turn_tools);
-                let violation_err: Option<String> = if tool_violations.is_empty() {
-                    None
-                } else {
-                    Some(format!(
-                        "[VALIDATION ERROR] Tool isolation violation: agent used disallowed tools: [{}]. Only [{}] are permitted.",
-                        tool_violations.join(", "),
-                        turn_tools.join(", ")
-                    ))
-                };
-                let hook_err = {
-                    let modified = helpers::detect_modified_files(project).await;
-                    if modified.is_empty() {
-                        None
-                    } else {
-                        let hook_event = harness_core::interceptor::ToolUseEvent {
-                            tool_name: "file_write".to_string(),
-                            affected_files: modified,
-                            session_id: None,
-                        };
-                        helpers::run_post_tool_use(interceptors, &hook_event, project).await
-                    }
-                };
-                let post_err = helpers::run_post_execute(interceptors, &turn_req, response).await;
-                if let Some(err) = violation_err.or(hook_err).or(post_err) {
-                    if validation_attempt < max_validation_retries {
-                        validation_attempt += 1;
-                        let backoff_ms = implement_pipeline::compute_backoff_ms(
-                            req.retry_base_backoff_ms,
-                            req.retry_max_backoff_ms,
-                            validation_attempt,
-                        );
-                        let truncated = helpers::truncate_validation_error(&err, 2000);
-                        turn_req.prompt = prompts::validation_retry_prompt(
-                            &first_req.prompt,
-                            validation_attempt,
-                            max_validation_retries,
-                            &truncated,
-                        );
-                        sleep(Duration::from_millis(backoff_ms)).await;
-                        continue;
-                    }
-                    helpers::run_on_error(interceptors, &turn_req, &err).await;
-                    anyhow::bail!(
-                        "Post-execution validation failed after {} attempts: {}",
-                        max_validation_retries,
-                        err
-                    );
-                }
-                break success;
-            }
-            Ok(Err(err)) => {
-                helpers::run_on_error(interceptors, &turn_req, &err.error.to_string()).await;
-                return Err(err.error.into());
-            }
-            Err(_) => {
-                let msg = format!(
-                    "{} timed out after {}s",
-                    task_kind.as_ref(),
-                    turn_timeout.as_secs()
-                );
-                helpers::run_on_error(interceptors, &turn_req, &msg).await;
-                anyhow::bail!(msg);
-            }
-        }
-    };
-
-    if implement_pipeline::contains_worktree_collision_sentinel(&resp.response.output) {
-        let collision_telemetry = resp.telemetry.clone();
-        mutate_and_persist(store, task_id, |s| {
-            s.status = TaskStatus::Failed;
-            s.turn = 1;
-            s.error = Some(
-                "WorktreeCollision: agent observed worktree managed by another harness session"
-                    .into(),
-            );
-            s.rounds.push(crate::task_runner::RoundResult::new(
-                1,
-                task_kind.as_ref().to_string(),
-                "worktree_collision",
-                if resp.response.output.is_empty() {
-                    None
-                } else {
-                    Some(resp.response.output.clone())
-                },
-                Some(collision_telemetry),
-                None,
-            ));
-        })
-        .await?;
-        tracing::info!(
-            task_id = %task_id,
-            task_kind = task_kind.as_ref(),
-            status = "failed",
-            total_elapsed_secs = task_start.elapsed().as_secs(),
-            "task_completed"
-        );
-        return Ok(());
-    }
-
-    let done_telemetry = resp.telemetry.clone();
-    mutate_and_persist(store, task_id, |s| {
-        s.status = TaskStatus::Done;
-        s.turn = 1;
-        s.rounds.push(crate::task_runner::RoundResult::new(
-            1,
-            task_kind.as_ref().to_string(),
-            "completed",
-            if resp.response.output.is_empty() {
-                None
-            } else {
-                Some(resp.response.output.clone())
-            },
-            Some(done_telemetry),
-            None,
-        ));
-    })
-    .await?;
-    store.log_event(crate::event_replay::TaskEvent::Completed {
-        task_id: task_id.0.clone(),
-        ts: crate::event_replay::now_ts(),
-    });
-    let event_name = format!("task_{}", task_kind.as_ref());
-    let mut ev = harness_core::types::Event::new(
-        harness_core::types::SessionId::new(),
-        &event_name,
-        "task_runner",
-        harness_core::types::Decision::Complete,
-    );
-    ev.detail = Some(format!("task_id={}", task_id.as_str()));
-    if let Err(err) = events.log(&ev).await {
-        tracing::warn!("failed to log {} event: {err}", task_kind.as_ref());
-    }
-    tracing::info!(
-        task_id = %task_id,
-        task_kind = task_kind.as_ref(),
-        status = "done",
-        total_elapsed_secs = task_start.elapsed().as_secs(),
-        "task_completed"
-    );
-    Ok(())
-}
-
-enum ConflictGateOutcome {
-    Clean,
-    RebasePushed,
-    Failed,
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn run_resumed_pr_conflict_gate(
-    store: &TaskStore,
-    task_id: &TaskId,
-    agent: &dyn CodeAgent,
-    req: &CreateTaskRequest,
-    project: &std::path::Path,
-    pr_num: u64,
-    repo_slug: &str,
-    context_items: &[harness_core::types::ContextItem],
-    interceptors: &Arc<Vec<Arc<dyn harness_core::interceptor::TurnInterceptor>>>,
-    events: &Arc<harness_observe::event_store::EventStore>,
-    cargo_env: &HashMap<String, String>,
-    turn_timeout: Duration,
-    effective_max_turns: Option<u32>,
-    turns_used: &mut u32,
-    turns_used_acc: &mut u32,
-) -> anyhow::Result<ConflictGateOutcome> {
-    use crate::task_executor::conflict_resolver::{
-        parse_conflict_check_output, ConflictCheckOutcome,
-    };
-
-    let gate_turn = turns_used.saturating_add(1);
-    let persist_failure =
-        |result: &'static str,
-         reason: String,
-         detail: Option<String>,
-         telemetry: Option<harness_core::types::TurnTelemetry>,
-         failure: Option<harness_core::types::TurnFailure>| async move {
-            mutate_and_persist(store, task_id, |s| {
-                s.status = TaskStatus::Failed;
-                s.turn = gate_turn;
-                s.error = Some(reason.clone());
-                s.rounds.push(crate::task_runner::RoundResult::new(
-                    gate_turn,
-                    "conflict_gate",
-                    result,
-                    detail.clone(),
-                    telemetry.clone(),
-                    failure.clone(),
-                ));
-            })
-            .await?;
-            let event = helpers::build_task_event(
-                task_id,
-                gate_turn,
-                "conflict_gate",
-                "pr_conflict_gate",
-                harness_core::types::Decision::Block,
-                Some(reason),
-                Some(format!("pr={pr_num}")),
-                telemetry,
-                failure,
-                detail,
-            );
-            if let Err(error) = events.log(&event).await {
-                tracing::warn!("failed to log pr_conflict_gate event: {error}");
-            }
-            Ok::<ConflictGateOutcome, anyhow::Error>(ConflictGateOutcome::Failed)
-        };
-
-    if let Some(max) = effective_max_turns {
-        if *turns_used >= max {
-            return persist_failure(
-                "turn_budget_exhausted",
-                format!(
-                    "pr:{pr_num} conflict gate could not run before review; manual resolution required: turn budget exhausted after {} of {} allowed turns",
-                    turns_used, max
-                ),
-                None,
-                None,
-                None,
-            )
-            .await;
-        }
-    }
-
-    let prompt_built_at = Utc::now();
-    let gate_prompt = prompts::check_resumed_pr_conflicts(pr_num, repo_slug, project);
-    let gate_req = AgentRequest {
-        prompt: gate_prompt,
-        project_root: project.to_path_buf(),
-        context: context_items.to_vec(),
-        max_budget_usd: req.max_budget_usd,
-        execution_phase: Some(harness_core::types::ExecutionPhase::Execution),
-        env_vars: cargo_env.clone(),
-        ..Default::default()
-    };
-    let gate_req = helpers::run_pre_execute(interceptors, gate_req).await?;
-    let gate_started_at = Utc::now();
-    let gate_resp = tokio::time::timeout(
-        turn_timeout,
-        helpers::run_agent_streaming(
-            agent,
-            gate_req.clone(),
-            task_id,
-            store,
-            gate_turn,
-            prompt_built_at,
-            gate_started_at,
-        ),
-    )
-    .await;
-    *turns_used += 1;
-    *turns_used_acc = *turns_used;
-
-    let (response, telemetry) = match gate_resp {
-        Ok(Ok(success)) => {
-            let response = success.response;
-            if let Some(validation_err) =
-                helpers::run_post_execute(interceptors, &gate_req, &response).await
-            {
-                helpers::run_on_error(interceptors, &gate_req, &validation_err).await;
-                return persist_failure(
-                    "validation_failed",
-                    format!(
-                        "pr:{pr_num} conflict gate failed closed; manual resolution required: post-execution validation failed: {validation_err}"
-                    ),
-                    if response.output.is_empty() {
-                        None
-                    } else {
-                        Some(response.output)
-                    },
-                    Some(success.telemetry),
-                    None,
-                )
-                .await;
-            }
-            (response, success.telemetry)
-        }
-        Ok(Err(failure)) => {
-            helpers::run_on_error(interceptors, &gate_req, &failure.error.to_string()).await;
-            return persist_failure(
-                "failed",
-                format!(
-                    "pr:{pr_num} conflict gate failed closed; manual resolution required: {}",
-                    failure.error
-                ),
-                None,
-                Some(failure.telemetry),
-                Some(failure.failure),
-            )
-            .await;
-        }
-        Err(_) => {
-            let msg = format!(
-                "pr:{pr_num} conflict gate timed out; manual resolution required after {}s",
-                turn_timeout.as_secs()
-            );
-            helpers::run_on_error(interceptors, &gate_req, &msg).await;
-            let telemetry =
-                helpers::telemetry_for_timeout(prompt_built_at, gate_started_at, Utc::now(), None);
-            let failure = harness_core::types::TurnFailure {
-                kind: harness_core::types::TurnFailureKind::Timeout,
-                provider: Some(agent.name().to_string()),
-                upstream_status: None,
-                message: Some(msg.clone()),
-                body_excerpt: None,
-            };
-            return persist_failure("timeout", msg, None, Some(telemetry), Some(failure)).await;
-        }
-    };
-
-    let detail = if response.output.is_empty() {
-        None
-    } else {
-        Some(response.output.clone())
-    };
-    match parse_conflict_check_output(&response.output) {
-        Ok(ConflictCheckOutcome::CleanPr) => {
-            mutate_and_persist(store, task_id, |s| {
-                s.rounds.push(crate::task_runner::RoundResult::new(
-                    gate_turn,
-                    "conflict_gate",
-                    "clean",
-                    detail.clone(),
-                    Some(telemetry.clone()),
-                    None,
-                ));
-            })
-            .await?;
-            let event = helpers::build_task_event(
-                task_id,
-                gate_turn,
-                "conflict_gate",
-                "pr_conflict_gate",
-                harness_core::types::Decision::Complete,
-                Some("resumed PR conflict gate passed cleanly".to_string()),
-                Some(format!("pr={pr_num}")),
-                Some(telemetry),
-                None,
-                detail,
-            );
-            if let Err(error) = events.log(&event).await {
-                tracing::warn!("failed to log pr_conflict_gate event: {error}");
-            }
-            Ok(ConflictGateOutcome::Clean)
-        }
-        Ok(ConflictCheckOutcome::RebasePushed) => {
-            mutate_and_persist(store, task_id, |s| {
-                s.rounds.push(crate::task_runner::RoundResult::new(
-                    gate_turn,
-                    "conflict_gate",
-                    "rebase_pushed",
-                    detail.clone(),
-                    Some(telemetry.clone()),
-                    None,
-                ));
-            })
-            .await?;
-            let event = helpers::build_task_event(
-                task_id,
-                gate_turn,
-                "conflict_gate",
-                "pr_conflict_gate",
-                harness_core::types::Decision::Complete,
-                Some("resumed PR conflict gate rebased and pushed".to_string()),
-                Some(format!("pr={pr_num}")),
-                Some(telemetry),
-                None,
-                detail,
-            );
-            if let Err(error) = events.log(&event).await {
-                tracing::warn!("failed to log pr_conflict_gate event: {error}");
-            }
-            Ok(ConflictGateOutcome::RebasePushed)
-        }
-        Ok(ConflictCheckOutcome::ManualResolutionRequired) => {
-            persist_failure(
-                "manual_resolution_required",
-                format!(
-                    "pr:{pr_num} is conflicting and rebase was not pushed; manual resolution required"
-                ),
-                detail,
-                Some(telemetry),
-                None,
-            )
-            .await
-        }
-        Err(parse_err) => {
-            persist_failure(
-                "malformed_output",
-                format!(
-                    "pr:{pr_num} conflict gate failed closed; manual resolution required: malformed output ({parse_err})"
-                ),
-                detail,
-                Some(telemetry),
-                None,
-            )
-            .await
-        }
-    }
-}
-
 pub(crate) async fn run_task(
     store: &TaskStore,
     task_id: &TaskId,
@@ -819,7 +227,7 @@ pub(crate) async fn run_task(
 
     if matches!(task_kind, TaskKind::Review | TaskKind::Planner) {
         let mut turns_used = *turns_used_acc;
-        run_non_implementation_task(
+        non_implementation::run_non_implementation_task(
             store,
             task_id,
             task_kind,
@@ -876,11 +284,11 @@ pub(crate) async fn run_task(
     };
     let resumed_review_prep = task_state
         .as_ref()
-        .and_then(|task| review_prep_from_rounds(&task.rounds))
+        .and_then(|task| gates::review_prep_from_rounds(&task.rounds))
         .or_else(|| {
             checkpoint
                 .as_ref()
-                .and_then(|c| review_prep_from_checkpoint_phase(&c.last_phase))
+                .and_then(|c| gates::review_prep_from_checkpoint_phase(&c.last_phase))
         });
     let resumed_pr_url: Option<String> =
         task_pr_url.or_else(|| checkpoint.as_ref().and_then(|c| c.pr_url.clone()));
@@ -1157,10 +565,10 @@ pub(crate) async fn run_task(
         return Ok(());
     }
 
-    let mut rebase_pushed = match review_entry_decision(pr_num, review_prep.as_ref()) {
-        ReviewEntryDecision::Proceed { rebase_pushed } => rebase_pushed,
-        ReviewEntryDecision::FailConflict { error, paths_csv } => {
-            fail_rebase_conflict(store, task_id, &events, pr_num, error, paths_csv).await?;
+    let mut rebase_pushed = match gates::review_entry_decision(pr_num, review_prep.as_ref()) {
+        gates::ReviewEntryDecision::Proceed { rebase_pushed } => rebase_pushed,
+        gates::ReviewEntryDecision::FailConflict { error, paths_csv } => {
+            gates::fail_rebase_conflict(store, task_id, &events, pr_num, error, paths_csv).await?;
             tracing::info!(
                 task_id = %task_id,
                 status = "failed",
@@ -1174,7 +582,7 @@ pub(crate) async fn run_task(
     };
 
     if review_prep.is_none() && was_resumed_pr {
-        match run_resumed_pr_conflict_gate(
+        match gates::run_resumed_pr_conflict_gate(
             store,
             task_id,
             agent,
@@ -1193,13 +601,13 @@ pub(crate) async fn run_task(
         )
         .await?
         {
-            ConflictGateOutcome::Clean => {
+            gates::ConflictGateOutcome::Clean => {
                 rebase_pushed = false;
             }
-            ConflictGateOutcome::RebasePushed => {
+            gates::ConflictGateOutcome::RebasePushed => {
                 rebase_pushed = true;
             }
-            ConflictGateOutcome::Failed => return Ok(()),
+            gates::ConflictGateOutcome::Failed => return Ok(()),
         }
     }
 
@@ -1313,440 +721,5 @@ pub(crate) async fn run_task(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn periodic_review_source_uses_standard_allowed_tools() {
-        // Verifies that periodic_review tasks get a non-empty allowed_tools list,
-        // which causes claude.rs to pass --allowedTools (hard enforcement) instead
-        // of --dangerously-skip-permissions.
-        let tools = restricted_tools(CapabilityProfile::Standard).unwrap_or_default();
-        assert_eq!(
-            tools,
-            CapabilityProfile::Standard.tools().unwrap_or_default()
-        );
-        assert!(!tools.is_empty());
-    }
-
-    #[test]
-    fn standard_implementation_turn_uses_full_profile() {
-        // Non-periodic_review tasks use None → Full profile →
-        // --dangerously-skip-permissions in claude.rs.
-        let implementation_allowed_tools: Option<Vec<String>> = None;
-        assert!(implementation_allowed_tools.is_none());
-    }
-
-    #[test]
-    fn parse_harness_review_command() {
-        let cmd = parse_harness_mention_command("@harness review");
-        assert_eq!(cmd, Some(HarnessMentionCommand::Review));
-    }
-
-    #[test]
-    fn parse_harness_fix_ci_command_case_insensitive() {
-        let cmd = parse_harness_mention_command("please @Harness FIX CI");
-        assert_eq!(cmd, Some(HarnessMentionCommand::FixCi));
-    }
-
-    #[test]
-    fn parse_harness_plain_mention_command() {
-        let cmd = parse_harness_mention_command("hello @harness can you help?");
-        assert_eq!(cmd, Some(HarnessMentionCommand::Mention));
-    }
-
-    #[test]
-    fn parse_harness_command_returns_none_without_mention() {
-        let cmd = parse_harness_mention_command("no command here");
-        assert_eq!(cmd, None);
-    }
-
-    #[test]
-    fn parse_harness_first_mention_per_line_is_used() {
-        let cmd = parse_harness_mention_command("@harness review then @harness fix ci");
-        assert_eq!(cmd, Some(HarnessMentionCommand::Review));
-    }
-
-    #[test]
-    fn prompt_builder_no_sections_adds_trailing_newline() {
-        let result = PromptBuilder::new("Title line.").build();
-        assert_eq!(result, "Title line.\n");
-    }
-
-    #[test]
-    fn prompt_builder_optional_url_absent_is_skipped() {
-        let result = PromptBuilder::new("Title.")
-            .add_optional_url("Link", None)
-            .build();
-        assert_eq!(result, "Title.\n");
-    }
-
-    #[test]
-    fn prompt_builder_optional_url_present_appears_in_output() {
-        let result = PromptBuilder::new("Title.")
-            .add_optional_url("Link", Some("https://example.com"))
-            .build();
-        assert!(result.contains("- Link: "));
-        assert!(result.contains("https://example.com"));
-        assert!(result.ends_with('\n'));
-    }
-
-    #[test]
-    fn prompt_builder_add_section_wraps_external_data() {
-        let result = PromptBuilder::new("Title.")
-            .add_section("Payload", "content here")
-            .build();
-        assert!(result.contains("Payload:\n"));
-        assert!(result.contains("<external_data>"));
-        assert!(result.contains("content here"));
-    }
-
-    #[test]
-    fn prompt_builder_multiple_urls_all_appear() {
-        let result = PromptBuilder::new("Title.")
-            .add_optional_url("First", Some("url1"))
-            .add_optional_url("Second", None)
-            .add_optional_url("Third", Some("url3"))
-            .build();
-        assert!(result.contains("- First: "));
-        assert!(result.contains("url1"));
-        assert!(!result.contains("Second"));
-        assert!(result.contains("- Third: "));
-        assert!(result.contains("url3"));
-    }
-
-    #[test]
-    fn build_fix_ci_prompt_contains_context() {
-        let prompt = build_fix_ci_prompt(
-            "majiayu000/harness",
-            42,
-            "@harness fix CI",
-            Some("https://github.com/majiayu000/harness/issues/42#issuecomment-1"),
-            Some("https://github.com/majiayu000/harness/pull/42"),
-        );
-
-        assert!(prompt.contains("CI failure repair requested for PR #42"));
-        assert!(prompt.contains("majiayu000/harness"));
-        assert!(prompt.contains("<external_data>"));
-        assert!(prompt.contains("PR_URL=https://github.com/majiayu000/harness/pull/42"));
-    }
-
-    #[test]
-    fn truncate_short_string_passes_through() {
-        let input = "short error";
-        let result = helpers::truncate_validation_error(input, 100);
-        assert_eq!(result, "short error");
-    }
-
-    #[test]
-    fn truncate_at_max_chars_boundary() {
-        let input = "a".repeat(200);
-        let result = helpers::truncate_validation_error(&input, 50);
-        assert!(result.starts_with(&"a".repeat(50)));
-        assert!(result.contains("(output truncated, 200 chars total)"));
-    }
-
-    #[test]
-    fn truncate_preserves_utf8_boundary() {
-        // "é" is 2 bytes; build a string where max_chars lands mid-character.
-        let input = "ééééé"; // 10 bytes, 5 chars
-        let result = helpers::truncate_validation_error(input, 3); // byte 3 is mid-char
-                                                                   // Should back up to byte 2 (1 full "é").
-        assert!(result.starts_with("é"));
-        assert!(result.contains("(output truncated,"));
-    }
-
-    #[test]
-    fn review_check_turn_uses_readonly_profile() {
-        let tools = restricted_tools(CapabilityProfile::ReadOnly).unwrap();
-        assert!(tools.contains(&"Read".to_string()));
-        assert!(tools.contains(&"Grep".to_string()));
-        assert!(tools.contains(&"Glob".to_string()));
-        assert!(!tools.contains(&"Write".to_string()));
-        assert!(!tools.contains(&"Edit".to_string()));
-        assert!(!tools.contains(&"Bash".to_string()));
-    }
-
-    #[test]
-    fn periodic_review_turn_uses_standard_profile_with_bash() {
-        let tools = restricted_tools(CapabilityProfile::Standard).unwrap();
-        assert!(tools.contains(&"Bash".to_string()));
-        assert!(tools.contains(&"Read".to_string()));
-        assert!(tools.contains(&"Write".to_string()));
-        assert!(tools.contains(&"Edit".to_string()));
-        // Standard does not include Grep/Glob — it's distinct from ReadOnly.
-        assert!(!tools.contains(&"Grep".to_string()));
-    }
-
-    #[test]
-    fn implementation_turn_uses_full_profile_no_restriction() {
-        // Full profile returns None — no tool restriction is applied to the agent.
-        assert!(CapabilityProfile::Full.tools().is_none());
-    }
-
-    // --- Gate: task_needs_pr_url covers issue and pr:N tasks ---
-
-    #[test]
-    fn task_needs_pr_url_true_for_issue_task() {
-        let req = CreateTaskRequest {
-            issue: Some(42),
-            ..CreateTaskRequest::default()
-        };
-        assert!(
-            implement_pipeline::task_needs_pr_url(&req),
-            "issue task must require PR_URL"
-        );
-    }
-
-    #[test]
-    fn issue_triage_runs_only_when_not_skipped_and_no_existing_pr() {
-        assert!(should_run_issue_triage(false, false));
-        assert!(!should_run_issue_triage(true, false));
-        assert!(!should_run_issue_triage(false, true));
-    }
-
-    #[test]
-    fn task_needs_pr_url_true_for_pr_task() {
-        let req = CreateTaskRequest {
-            pr: Some(99),
-            ..CreateTaskRequest::default()
-        };
-        assert!(
-            implement_pipeline::task_needs_pr_url(&req),
-            "pr:N task must require PR_URL"
-        );
-    }
-
-    #[test]
-    fn task_needs_pr_url_false_for_prompt_only_task() {
-        let req = CreateTaskRequest::default();
-        assert!(
-            !implement_pipeline::task_needs_pr_url(&req),
-            "prompt-only task must not require PR_URL (Done is correct)"
-        );
-    }
-
-    #[test]
-    fn review_entry_decision_blocks_review_on_rebase_conflict() {
-        assert_eq!(
-            review_entry_decision(
-                42,
-                Some(&prompts::PrReviewPrepOutcome::RebaseConflict {
-                    paths: vec!["src/lib.rs".to_string(), "src/main.rs".to_string()],
-                })
-            ),
-            ReviewEntryDecision::FailConflict {
-                error: "PR #42 has rebase conflicts: src/lib.rs, src/main.rs; manual resolution required".to_string(),
-                paths_csv: "src/lib.rs, src/main.rs".to_string(),
-            }
-        );
-    }
-
-    #[test]
-    fn review_entry_decision_propagates_rebase_pushed_flag() {
-        assert_eq!(
-            review_entry_decision(42, Some(&prompts::PrReviewPrepOutcome::RebasePushed)),
-            ReviewEntryDecision::Proceed {
-                rebase_pushed: true
-            }
-        );
-        assert_eq!(
-            review_entry_decision(42, Some(&prompts::PrReviewPrepOutcome::RebaseSkipped)),
-            ReviewEntryDecision::Proceed {
-                rebase_pushed: false
-            }
-        );
-    }
-
-    #[test]
-    fn review_prep_from_rounds_prefers_latest_implement_round() {
-        let rounds = vec![
-            RoundResult::new(1, "implement", "rebase_skipped", None, None, None),
-            RoundResult::new(2, "review", "needs_fix", None, None, None),
-            RoundResult::new(3, "implement", "rebase_pushed", None, None, None),
-        ];
-        assert_eq!(
-            review_prep_from_rounds(&rounds),
-            Some(prompts::PrReviewPrepOutcome::RebasePushed)
-        );
-    }
-
-    #[test]
-    fn review_prep_from_rounds_recovers_rebase_conflict_detail() {
-        let rounds = vec![RoundResult::new(
-            1,
-            "implement",
-            "rebase_conflict",
-            Some("REBASE_CONFLICT paths=src/lib.rs,src/main.rs".to_string()),
-            None,
-            None,
-        )];
-        assert_eq!(
-            review_prep_from_rounds(&rounds),
-            Some(prompts::PrReviewPrepOutcome::RebaseConflict {
-                paths: vec!["src/lib.rs".to_string(), "src/main.rs".to_string()],
-            })
-        );
-    }
-
-    #[test]
-    fn review_prep_from_checkpoint_phase_recovers_rebase_state() {
-        assert_eq!(
-            review_prep_from_checkpoint_phase("rebase_pushed"),
-            Some(prompts::PrReviewPrepOutcome::RebasePushed)
-        );
-        assert_eq!(
-            review_prep_from_checkpoint_phase("rebase_skipped"),
-            Some(prompts::PrReviewPrepOutcome::RebaseSkipped)
-        );
-        assert_eq!(
-            review_prep_from_checkpoint_phase("rebase_conflict"),
-            Some(prompts::PrReviewPrepOutcome::RebaseConflict { paths: Vec::new() })
-        );
-        assert_eq!(review_prep_from_checkpoint_phase("pr_created"), None);
-    }
-
-    #[tokio::test]
-    async fn fail_rebase_conflict_marks_task_failed_and_logs_event() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let store = TaskStore::open(&dir.path().join("tasks.db")).await?;
-        let events = Arc::new(harness_observe::event_store::EventStore::new(dir.path()).await?);
-        let task_id = TaskId::new();
-        let state = crate::task_runner::TaskState::new(task_id.clone());
-        store.insert(&state).await;
-
-        fail_rebase_conflict(
-            &store,
-            &task_id,
-            &events,
-            42,
-            "PR #42 has rebase conflicts: src/lib.rs, src/main.rs; manual resolution required"
-                .to_string(),
-            "src/lib.rs, src/main.rs".to_string(),
-        )
-        .await?;
-
-        let final_state = store
-            .get(&task_id)
-            .ok_or_else(|| anyhow::anyhow!("task must exist"))?;
-        assert!(matches!(final_state.status, TaskStatus::Failed));
-        assert_eq!(
-            final_state.error.as_deref(),
-            Some(
-                "PR #42 has rebase conflicts: src/lib.rs, src/main.rs; manual resolution required"
-            )
-        );
-
-        let logged = events
-            .query(&harness_core::types::EventFilters {
-                hook: Some("pr_rebase_conflict".to_string()),
-                ..harness_core::types::EventFilters::default()
-            })
-            .await?;
-        assert_eq!(logged.len(), 1);
-        assert_eq!(logged[0].reason.as_deref(), Some("src/lib.rs, src/main.rs"));
-        let expected_detail = format!("task_id={} pr=42", task_id.as_str());
-        assert_eq!(logged[0].detail.as_deref(), Some(expected_detail.as_str()));
-        Ok(())
-    }
-
-    // --- Gate A: pr:N task with non-empty output but no PR_URL gets Failed ---
-
-    #[tokio::test]
-    async fn pr_task_nonempty_output_no_pr_url_marks_failed() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let store = TaskStore::open(&dir.path().join("tasks.db")).await?;
-        let task_id = TaskId::new();
-        let state = crate::task_runner::TaskState::new(task_id.clone());
-        store.insert(&state).await;
-
-        let req = CreateTaskRequest {
-            pr: Some(42),
-            ..CreateTaskRequest::default()
-        };
-        // Non-empty output from agent but no PR_URL found — gate must fire for pr:N.
-        let output = "LGTM, nothing to change";
-        if output.trim().is_empty() {
-            mutate_and_persist(&store, &task_id, |s| {
-                s.status = TaskStatus::Failed;
-                s.turn = 2;
-                s.error = Some("empty agent output: no PR created and no output".to_string());
-            })
-            .await?;
-        } else if implement_pipeline::task_needs_pr_url(&req) {
-            mutate_and_persist(&store, &task_id, |s| {
-                s.status = TaskStatus::Failed;
-                s.turn = 2;
-                s.error =
-                    Some("no PR number found in agent output; task requires PR_URL".to_string());
-            })
-            .await?;
-        } else {
-            mutate_and_persist(&store, &task_id, |s| {
-                s.status = TaskStatus::Done;
-                s.turn = 2;
-            })
-            .await?;
-        }
-
-        let final_state = store
-            .get(&task_id)
-            .ok_or_else(|| anyhow::anyhow!("task must exist"))?;
-        assert!(
-            matches!(final_state.status, TaskStatus::Failed),
-            "pr:N task with non-empty output but no PR_URL must be Failed, not Done"
-        );
-        assert!(
-            final_state
-                .error
-                .as_deref()
-                .unwrap_or("")
-                .contains("PR_URL"),
-            "error must mention PR_URL"
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn prompt_only_nonempty_output_no_pr_stays_done() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let store = TaskStore::open(&dir.path().join("tasks.db")).await?;
-        let task_id = TaskId::new();
-        let state = crate::task_runner::TaskState::new(task_id.clone());
-        store.insert(&state).await;
-
-        let req = CreateTaskRequest::default(); // no issue, no pr
-        let output = "periodic review complete: no issues found";
-        if output.trim().is_empty() {
-            mutate_and_persist(&store, &task_id, |s| {
-                s.status = TaskStatus::Failed;
-                s.turn = 2;
-                s.error = Some("empty agent output: no PR created and no output".to_string());
-            })
-            .await?;
-        } else if implement_pipeline::task_needs_pr_url(&req) {
-            mutate_and_persist(&store, &task_id, |s| {
-                s.status = TaskStatus::Failed;
-                s.turn = 2;
-                s.error =
-                    Some("no PR number found in agent output; task requires PR_URL".to_string());
-            })
-            .await?;
-        } else {
-            mutate_and_persist(&store, &task_id, |s| {
-                s.status = TaskStatus::Done;
-                s.turn = 2;
-            })
-            .await?;
-        }
-
-        let final_state = store
-            .get(&task_id)
-            .ok_or_else(|| anyhow::anyhow!("task must exist"))?;
-        assert!(
-            matches!(final_state.status, TaskStatus::Done),
-            "prompt-only task with non-empty output and no PR must be Done"
-        );
-        Ok(())
-    }
-}
+#[path = "mod_tests.rs"]
+mod tests;
